First visualization
Added DOTS ECS logic for data conversion to visible entities, using manual entity creation, instead of the automatic conversion to entity workflow.
Added 7 individual star materials (instead of a function of star.BV)
Changed Star class double fields to float

+++ 3.Visualization/HipparcosVisualization/Assets/Scripts/Star.cs
@@ -4,13 +4,13 @@
    {
       public int Hip { get; set; }
       public int Nc { get; set; }
-      public double Rarad { get; set; }
-      public double Derad { get; set; }
-      public double Plx { get; set; }
-      public double Bv { get; set; }
-      public double Dly { get; set; }
-      public double Xly { get; set; }
-      public double Yly { get; set; }
-      public double Zly { get; set; }
+      public float Rarad { get; set; }
+      public float Derad { get; set; }
+      public float Plx { get; set; }
+      public float Bv { get; set; }
+      public float Dly { get; set; }
+      public float Xly { get; set; }
+      public float Yly { get; set; }
+      public float Zly { get; set; }
    }
 }
